удалены console.log, добавлен const.js, переименован data.js
