+++ 2672189-kekstagram-6/js/const.js
@@ -1,0 +1,38 @@
+const PHOTO_COUNT = 25;
+
+const NAMES = [
+  'Анна',
+  'Дима',
+  'Катя',
+  'Илья',
+  'Марина',
+  'Олег',
+  'София',
+  'Никита',
+  'Вика',
+  'Кирилл'
+];
+
+const DESCRIPTIONS = [
+  'Закат на море 🌅',
+  'Утренний кофе ☕',
+  'Прогулка по городу 🚶‍♀️',
+  'Лучший день! 😍',
+  'Пикник с друзьями 🧺',
+  'Ночь в горах 🌌',
+  'Любимый кот 🐈',
+  'Завтрак на террасе 🥐',
+  'Выходные с друзьями 🎉',
+  'Просто чудесный день 🌞'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делали эту фотографию, у меня упал телефон от зависти 😆',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Лайк за позитив 🔥',
+  'Шикарное фото! 💕'
+];
+
+export { PHOTO_COUNT, NAMES, DESCRIPTIONS, MESSAGES };
