+++ 2672189-kekstagram-6/js/utils.js
@@ -1,8 +1,7 @@
 const getRandomInteger = (a, b) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
+  return Math.floor(Math.random() * (upper - lower + 1)) + lower;
 };
 
 const getRandomArrayElement = (elements) =>
